feat: build image local
reference: https://izumisy.work/entry/2022/03/28/140644

+++ entrypoint.sh
@@ -1,9 +1,14 @@
 #!/bin/sh
 
-set -e
-pymajor=${PYVERSION%%.*} pyminor=${PYVERSION#"${pymajor}."} pyminor=${pyminor%.*}
+print () {
+  if [ -z "$debug" ]; then
+    printf "\033[36m\033[1H$1\033[0m\033[2H"
+  else
+    printf "$1\n"
+  fi
+}
 
-cp "/wine64/drive_c/Python${pymajor}${pyminor}/vcruntime140.dll" /src
+set -e
 WORKDIR=${SRCDIR:-/src}
 
 cd "$WORKDIR"
@@ -11,35 +16,73 @@ cd "$WORKDIR"
 PYPI_URL=${PYPI_URL:-"https://pypi.python.org/"}
 PYPI_INDEX_URL=${PYPI_INDEX_URL:-"https://pypi.python.org/simple"}
 
-echo "PYTHON_VERSION = ${PYVERSION}"
-echo "PYINSTALLER_VERSION = ${PYINSTALLER_VERSION}"
+first=$1
+while :; do
+  case $1 in
+    -r|--requirements)
+      requirements=$2
+      shift; shift
+      break
+      ;;
+    --requierements=*)
+      requirements=${1#*=}
+      shift
+      break
+      ;;
+    --debug)
+      debug="yes"
+      set -x
+      shift
+      first=$1
+      ;;
+    --random-key)
+      if [ "$WINEARCH" = "win64" ]; then
+        export USEKEY=1
+        random_key=$(pwgen -s 16 1)
+      fi
+      shift
+      first=$1
+      ;;
+    *)
+      tmp=$1
+      shift
+      set -- $@ $tmp
+      [ "$first" = "$1" ] && break
+      ;;
+  esac
+done
 
-if [ -f requirements.txt ]; then
-  pip --cache-dir "${PIP_CACHE_DIR:=./pip-cache}" install -r requirements.txt
+clear
+[ -z "$debug" ] && printf '\033[2;16r\033[2H'
+if [ -f "${requirements:=requirements.txt}" ]; then
+  print "pip installation"
+  pip --cache-dir "${PIP_CACHE_DIR:=./pip-cache}" install -r ${requirements}
 fi
-
-echo "$@"
 
 for str in "$@"; do
   case $str in
     *.py)
+      print "compiler: pyinstaller, version: ${PYINSTALLER_VERSION}"
         pyinstaller \
           --clean --noconfirm \
           --onefile \
-          --distpath . \
+          --distpath ./dist \
           --workpath /tmp \
-          --runtime-tmpdir . \
           -p . \
-          $@ ;;
+          ${@/--random-key/--key $random_key}
+        break
+        ;;
     *.spec)
-      exename="${str%.spec}.exe"
+      print "compiler: pyinstaller, version: ${PYINSTALLER_VERSION} (using spec file)"
       pyinstaller \
         --clean --noconfirm \
-          --distpath . \
+          --distpath ./dist \
           --workpath /tmp \
-          $str ;;
+          $@
+      break
+      ;;
   esac
 done
+printf '\e[r\e[0m\e[17H'
 
-[ $RUN_EXE ] && wine "$exename" || true
-rm -r __pycache__
+[ -d "__pycache__" ] && rm -r __pycache__
